Moved common methods to package object kesque
