+++ kesque/src/main/scala/kesque/Kesque.scala
@@ -2,18 +2,11 @@ package kesque
 
 import kafka.server.QuotaFactory.UnboundedQuota
 import java.io.File
-import java.nio.ByteBuffer
 import java.util.Properties
 import org.apache.kafka.common.TopicPartition
-import org.apache.kafka.common.record.AbstractRecords
 import org.apache.kafka.common.record.CompressionType
-import org.apache.kafka.common.record.MemoryRecords
-import org.apache.kafka.common.record.MemoryRecordsBuilder
-import org.apache.kafka.common.record.RecordBatch
 import org.apache.kafka.common.record.SimpleRecord
-import org.apache.kafka.common.record.TimestampType
 import org.apache.kafka.common.requests.FetchRequest.PartitionData
-import org.apache.kafka.common.utils.ByteBufferOutputStream
 import scala.collection.mutable
 
 /**
@@ -65,38 +58,6 @@ object Kesque {
       println(value) // Some(value_1), Some(value_2), Some(value_3)
     }
   }
-
-  def buildRecords(compressionType: CompressionType, initialOffset: Long, records: SimpleRecord*): MemoryRecords = buildRecords(
-    RecordBatch.CURRENT_MAGIC_VALUE, initialOffset, compressionType,
-    TimestampType.CREATE_TIME, 0, 0,
-    0, RecordBatch.NO_PARTITION_LEADER_EPOCH, isTransactional = false,
-    records: _*
-  )
-
-  def buildRecords(magic: Byte, initialOffset: Long, compressionType: CompressionType,
-                   timestampType: TimestampType, producerId: Long, producerEpoch: Short,
-                   baseSequence: Int, partitionLeaderEpoch: Int, isTransactional: Boolean,
-                   records: SimpleRecord*): MemoryRecords = {
-    if (records.isEmpty) {
-      MemoryRecords.EMPTY
-    } else {
-      import scala.collection.JavaConverters._
-      val sizeEstimate = AbstractRecords.estimateSizeInBytes(magic, compressionType, records.asJava)
-      val bufferStream = new ByteBufferOutputStream(sizeEstimate)
-      val logAppendTime = timestampType match {
-        case TimestampType.LOG_APPEND_TIME => System.currentTimeMillis()
-        case _                             => RecordBatch.NO_TIMESTAMP
-      }
-
-      val builder = new MemoryRecordsBuilder(bufferStream, magic, compressionType, timestampType,
-        initialOffset, logAppendTime, producerId, producerEpoch, baseSequence, isTransactional, false,
-        partitionLeaderEpoch, sizeEstimate)
-
-      records foreach builder.append
-
-      builder.build()
-    }
-  }
 }
 final class Kesque(props: Properties) {
   private val kafkaServer = KafkaServer.start(props)
@@ -136,7 +97,7 @@ final class Kesque(props: Properties) {
     //val initialOffset = readerWriter.getLogEndOffset(partition) + 1 // TODO check -1L
     val initialOffset = 0L // TODO is this useful?
 
-    val memoryRecords = Kesque.buildRecords(compressionType, initialOffset, records: _*)
+    val memoryRecords = kesque.buildRecords(compressionType, initialOffset, records: _*)
     val entriesPerPartition = Map(partition -> memoryRecords)
 
     replicaManager.appendToLocalLog(
@@ -182,8 +143,8 @@ final class Kesque(props: Properties) {
     var lastOffset = fetchOffset
     while (recs.hasNext) {
       val rec = recs.next
-      val key = if (rec.hasKey) getBytes(rec.key) else null
-      val value = if (rec.hasValue) getBytes(rec.value) else null
+      val key = if (rec.hasKey) kesque.getBytes(rec.key) else null
+      val value = if (rec.hasValue) kesque.getBytes(rec.value) else null
       val timestamp = rec.timestamp
       val offset = rec.offset
       op(offset, TKeyVal(key, value, timestamp))
@@ -195,30 +156,9 @@ final class Kesque(props: Properties) {
     (i, lastOffset)
   }
 
-  /**
-   * Special function to extract bytes from kafka's DefaultRecord key or value ByteBuffer
-   * @see org.apache.kafka.common.utils.Utils.writeTo
-   */
-  private[kesque] def getBytes(buffer: ByteBuffer): Array[Byte] = {
-    val length = buffer.remaining
-    val value = Array.ofDim[Byte](length)
-    if (buffer.hasArray) {
-      System.arraycopy(buffer.array, buffer.position + buffer.arrayOffset, value, 0, length)
-    } else {
-      val pos = buffer.position
-      var i = pos
-      while (i < length + pos) {
-        value(i) = buffer.get(i)
-        i += 1
-      }
-    }
-    value
-  }
-
   def shutdown() {
     kafkaServer.shutdown()
   }
-
 }
 
 final case class TKeyVal(key: Array[Byte], value: Array[Byte], timestamp: Long = -1L)
